+++ Source/OCMockito/Mocking/MKTProtocolMock.m
@@ -56,6 +56,7 @@
 #pragma mark Support being called as isEqual: argument
 
 - (BOOL)isNSArray__ { return NO; }
+- (BOOL)isNSData__ { return NO; }
 - (BOOL)isNSDictionary__ { return NO; }
 - (BOOL)isNSNumber__ { return NO; }
 - (BOOL)isNSString__ { return NO; }